web/requirements.txt
Django==1.10
django-cors-headers==1.1.0
django-smtp-ssl==1.0
djangorestframework==3.4.3
djoser==0.5.0
drfdocs==0.0.11
Markdown==2.6.6
django-crispy-forms==1.6.0
django-filter==0.13.0
gunicorn==19.6.0
python-dateutil==2.5.3
setproctitle==1.1.10
wsgiref==0.1.2
psycopg2==2.6.2
whitenoise==3.2.1
